+++ CHANGELOG.md
@@ -219,6 +219,10 @@ None
   revisions by default (defined by `revsets.op-diff-changes-in`). A new flag,
   `--show-changes-in`, can be used to override this. [#6083](https://github.com/jj-vcs/jj/issues/6083)
 
+* `jj arrange` now supports a "squash" action for squashing into the parent and
+  keeping the parent's description.
+
+
 ### Fixed bugs
 
 * `.gitignore` with UTF-8 BOM can now be parsed correctly.